Move TargetDevice to common
